test(jwt): cancellation flag の component-level regression test を追加 (#389)
PR #432 レビュー指摘対応。`verifySignature` を `src/utils/jwt-verify.ts`
に切り出し、ESM mock 可能にした上で、stale promise の resolve が新しい
verify 結果を上書きしないことを assert する component-level unit test
を追加。

検証フロー:
1. secretKey='first' で verify 呼び出し（call 1）
2. secretKey='second' で verify 呼び出し（call 2、call 1 は cleanup で cancelled）
3. call 2 を 'invalid' で先に resolve → badge に '署名: 無効'
4. call 1 を 'valid' で後から resolve → cancellation flag により無視
5. badge が '署名: 無効' のまま変化しないことを assert

陽性対照: 同 test を旧実装 (cleanup なし) に当てて fail することを
ローカルで確認済み (call 1 の stale resolve が '署名: 有効' に上書き)。

既存の `JwtDecoder.test.ts` は import を `'@/utils/jwt-verify'` に
切替（コードベース全体の責務をより明確化）。
